OPENEULER: net: sfp: Synchronize some CMIS transceiver modules from ethtool
Mainline: KYLIN-only
Severity: Moderate

task: 218995

Currently, the SFF-8024 Identifier Values that according to the standard
support for the Common Management Interface Specification (CMIS)
based on standard identifier values in the ethtool is more than in the
kernel. When the driver needs to use a newer Identifier Value, the kernel
interface does not support it. Therefore, we synchronize the CMIS mode
Identifier Values which supported by the ethtool to the kernel.

Signed-off-by: Hao Lan <lanhao@huawei.com>
Signed-off-by: Jiantao Xiao <xiaojiantao1@h-partners.com>

K2CI-Arch: All
Link: https://gitee.com/openeuler/kernel/issues/I9A3QT
Signed-off-by: lifangxu <lifangxu@kylinos.cn>
Change-Id: I33c965637ba0019363eb1d7257a30567eca879a6
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/111993
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>