media: bttv: add back vbi hack
Mainline: 3f1faa154a4316b1b585a25394697504b3c24e98
From: v6.8-rc1
Severity: Moderate

task: 196396

[ Upstream commit 3f1faa154a4316b1b585a25394697504b3c24e98 ]

The old (now removed) videobuf framework had an optional vbi hack where
the sequence number of the frame counter was copied in the last 4 bytes
of the buffer. This hack was active only for the read() interface
(so not for streaming I/O), and it was enabled by bttv. This allowed
applications that used read() for the VBI data to match it with the
corresponding video frame.

When bttv was converted to vb2 this hack was forgotten, but some old
applications rely on this.

So add this back, but this time in the bttv driver rather than in the
vb2 framework.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Fixes: b7ec3212a73a ("media: bttv: convert to vb2")
Tested-by: Dr. David Alan Gilbert <dave@treblig.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 54ce93d7e66ee3e0451103ff2ff9a9b227f9f40a)

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Change-Id: I33c54e5a38e84d5053b4c2f01c8d165baa6c473f
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/94333
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>