cgroup/rstat: Optimize cgroup_rstat_updated_list()
Mainline: d499fd418fa15949d86d28bb5442ab88203fc513
From: v6.8-rc1
Severity: Low

task: 455682

The current design of cgroup_rstat_cpu_pop_updated() is to traverse
the updated tree in a way to pop out the leaf nodes first before
their parents. This can cause traversal of multiple nodes before a
leaf node can be found and popped out. IOW, a given node in the tree
can be visited multiple times before the whole operation is done. So
it is not very efficient and the code can be hard to read.

With the introduction of cgroup_rstat_updated_list() to build a list
of cgroups to be flushed first before any flushing operation is being
done, we can optimize the way the updated tree nodes are being popped
by pushing the parents first to the tail end of the list before their
children. In this way, most updated tree nodes will be visited only
once with the exception of the subtree root as we still need to go
back to its parent and popped it out of its updated_children list.
This also makes the code easier to read.

Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Tejun Heo <tj@kernel.org>

(backported from commit d499fd418fa15949d86d28bb5442ab88203fc513)

Signed-off-by: zhangguopeng <zhangguopeng@kylinos.cn>
Change-Id: Ia40502731908abdebc05dc76d3526213f8cb02a5
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/55847
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>