SUNWAY: fix lockdep error
Mainline: KYLIN-only
Severity: Low

Task: #382309

Initialize a dynamically allocated struct bin_attribute so we
can make lockdep happy.  This is a new requirement for
attributes and initially this is only needed when lockdep is
enabled.  Lockdep gives a nice error when your attribute is
added to sysfs if you don't have this.

K2CI-Arch: Sunway
Signed-off-by: He Chuyue <hechuyue@wxiat.com>
Signed-off-by: liuyunchen <liuyunchen@kylinos.cn>
Change-Id: Iaf6cc2e39bf3ba464d2e6bab796854621a0ea36e
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/53802
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>