OPENEULER: ipvlan: Introduce l2e mode
Mainline: KYLIN-only
Severity: Moderate

Task: #210115

In a typical IPvlan L2 setup where master is in default-ns and
each slave is into different (slave) ns. In this setup, if master
and slaves in different net, egress packet processing for traffic
originating from slave-ns can't be forwarded to master or other
machine whose ip in the same net with master, and they can't be
forwarded to other interface in default-ns.

This patch introuce a new mode l2e for ipvlan to realize above
goals, and it won't affect the original l2, l3, l3s mode.

As the ip tool doesn't support l2e mode, We use module param
"ipvlan_default_mode" to set the default work mode. 0 for l2
mode, 1 for l3, 2 for l2e, 3 for l3s, others invalid now.
Attention, when we create ipvlan devices by "ip" commond, if we
assign the mode, ipvlan will work in the mode we assigned other
then the "ipvlan_default_mode".

Signed-off-by: Keefe LIU <liuqifa@huawei.com>

Modified:
    when insmod the ipvlan module, the origin version(4.19) is 0
for l2, 1 for l3, 2 for l2e, 3 for l3s mode, but in 5.10 version
is 0 for l2, 1 for l3, 2 for l3s, 3 for l2e mode.

Signed-off-by: Lu Wei <luwei32@huawei.com>
Signed-off-by: Zhengchao Shao <shaozhengchao@huawei.com>

K2CI-Arch: All
Link: https://gitee.com/openeuler/kernel/issues/I8KDQE
Signed-off-by: Yongzhen Zhang <zhangyongzhen@kylinos.cn>
Change-Id: I88a780a6fca316468ca9d431455cade0a043f088
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/105594
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>