OPENEULER: ipvlan: Introduce local xmit queue for l2e mode
Mainline: KYLIN-only
Severity: Moderate

Task: #210115

Consider two IPVlan devices are set up on the same master, when
they communicate with each other by TCP, the receive part is too
fast to make the send packets coalesced, so in this case, the
performace is not as good as we expect.

This patch introduces a local xmit queue for l2e mode, when the
packets are sent to the IPVlan devices of the same master, the
packets will be cloned and added to the local xmit queue, this
operation can make the send packets coalesced and improve the
TCP performace in this case.

Signed-off-by: Keefe LIU <liuqifa@huawei.com>
Signed-off-by: Zhengchao Shao <shaozhengchao@huawei.com>

K2CI-Arch: All
Link: https://gitee.com/openeuler/kernel/issues/I8KDQE
Signed-off-by: Yongzhen Zhang <zhangyongzhen@kylinos.cn>
Change-Id: If3ca6f31c54298856bbfc234f75021dac803683e
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/105595
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>