bcachefs: return from fsync on writeback error to avoid early shutdown
Mainline: 5a11b5fe79e9f781db8aec6222d03dc92e2978d9
From: v6.8-rc1
Severity: Moderate

task: 205710

When investigating transient failures of generic/441 on bcachefs, it
was determined that the cause of the failure was a combination of
unconditional emergency shutdown and racing between background
journal activity and the test switchover from a working device
mapper table to an error injecting table.

Part of the reason for this sequence of events is that bcachefs
aggressively flushes as much as possible during fsync(), regardless
of errors. While this is reasonable behavior, it is technically
unnecessary because once an error is returned from fsync(), the
caller cannot make any assumptions about the resilience of data.

Tweak the bch2_fsync() logic to return an error on failure of any of
the steps involved in the flush. Note that this change alone does
not prevent generic/441 failure, but in combination with a test
tweak to avoid racing during the dm-error table switchover it avoids
the unnecessary shutdowns and allows the test to pass reliably on
bcachefs.

Signed-off-by: Brian Foster <bfoster@redhat.com>
Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>
(backported from commit 5a11b5fe79e9f781db8aec6222d03dc92e2978d9)

K2CI-Arch: All
Signed-off-by: Youling Tang <tangyouling@kylinos.cn>
Change-Id: Ie57d95daff3ad54fe6938d33c7e04a6b1c709435
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/107550
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>