cifs: reconnect worker should take reference on server struct unconditionally
Mainline: 04909192ada3285070f8ced0af7f07735478b364
From: v6.7-rc5
Severity: Moderate

task: 198402

[ Upstream commit 04909192ada3285070f8ced0af7f07735478b364 ]

Reconnect worker currently assumes that the server struct
is alive and only takes reference on the server if it needs
to call smb2_reconnect.

With the new ability to disable channels based on whether the
server has multichannel disabled, this becomes a problem when
we need to disable established channels. While disabling the
channels and deallocating the server, there could be reconnect
work that could not be cancelled (because it started).

This change forces the reconnect worker to unconditionally
take a reference on the server when it runs.

Also, this change now allows smb2_reconnect to know if it was
called by the reconnect worker. Based on this, the cifs_put_tcp_session
can decide whether it can cancel the reconnect work synchronously or not.

Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>
Signed-off-by: Steve French <stfrench@microsoft.com>
Stable-dep-of: 78e727e58e54 ("cifs: update iface_last_update on each query-and-update")
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 07e864089a0e4f521f4232b7f35d4723049c568b)

K2CI-Arch: None
Signed-off-by: k2ci <kernel-bot@kylinos.cn>
Change-Id: I44ce974b8d3b064ba0018461a8d6d091b7e3931a
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/96132
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>