OPENEULER: arm64: add support for machine check error safe
Mainline: KYLIN-only
Severity: Moderate

Task: #210128

For the arm64 kernel, when it processes hardware memory errors for
synchronize notifications(do_sea()), if the errors is consumed within the
kernel, the current processing is panic. However, it is not optimal.

Take uaccess for example, if the uaccess operation fails due to memory
error, only the user process will be affected. Killing the user process and
isolating the corrupt page is a better choice.

This patch only enable machine error check framework and adds an exception
fixup before the kernel panic in do_sea().

Signed-off-by: Tong Tiangen <tongtiangen@huawei.com>

K2CI-Arch: Arm64
Link: https://gitee.com/openeuler/kernel/issues/I8M74H
Signed-off-by: Yongzhen Zhang <zhangyongzhen@kylinos.cn>
Change-Id: I8c000bfb5df3e61673d70ac50195365b1fb92ca0
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/105777
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>