ACPI: utils: Fix error path in acpi_evaluate_reference()
Mainline: 8f0b960a42badda7a2781e8a33564624200debc9
From: v6.7-rc5
Severity: Low

Task: #551019

If a pointer to an uninitialized struct acpi_handle_list is passed to
acpi_evaluate_reference() and it decides to bail out early, either
because acpi_evaluate_object() fails, or because it produces invalid
data, the handles pointer from the struct acpi_handle_list will be
passed to kfree() and if it is not NULL, the kernel will crash on an
attempt to free unallocated memory.

Address this by moving the "end" label in acpi_evaluate_reference() to
the end of the function, which is sufficient, because no cleanup is
needed in that case.

Fixes: 2e57d10a6591 ("ACPI: utils: Dynamically determine acpi_handle_list size")
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Tested-by: Woody Suwalski <terraluna977@gmail.com>

K2CI-Arch: All
Change-Id: I9bc6b74eedaaf64de7c32f8724efa5abb7b24615
Signed-off-by: Yufeng Wang <wangyufeng@kylinos.cn>
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/146510
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>