drm/amd/display: Only clear symclk otg flag for HDMI
Mainline: dff45f03f508c92cd8eb2050e27b726726b8ae0b
From: v6.8-rc1
Severity: Moderate

task: 198520

[ Upstream commit dff45f03f508c92cd8eb2050e27b726726b8ae0b ]

[Description]
There is a corner case where the symclk otg flag is cleared
when disabling the phantom pipe for subvp (because the phantom
and main pipe share the same link). This is undesired because
we need the maintain the correct symclk otg flag state for
the main pipe.

For now only clear the flag only for HDMI signal type, since
it's only set for HDMI signal type (phantom is virtual). The
ideal solution is to not clear it if the stream is phantom but
currently there's a bug that doesn't allow us to do this. Once
this issue is fixed the proper fix can be implemented.

Reviewed-by: Samson Tam <samson.tam@amd.com>
Acked-by: Wayne Lin <wayne.lin@amd.com>
Signed-off-by: Alvin Lee <alvin.lee2@amd.com>
Tested-by: Daniel Wheeler <daniel.wheeler@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 49b4cab7bcbdfe621162405502283fc9cb3ef77f)

K2CI-Arch: None
Signed-off-by: k2ci <kernel-bot@kylinos.cn>
Change-Id: Iaf446a2f16f98b807d90f8149a7ad1b02cc8f62e
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/96692
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>