ACPI: utils: Return bool from acpi_evaluate_reference()
Mainline: 6909e0f322b0527fee9fdc54685e6cad69008713
From: v6.8-rc1
Severity: Low

Task: #551019

There are only 4 users of acpi_evaluate_reference() and none of them
actually cares about the reason why it fails.  All of them are only
interested in whether or not it is successful, so it can return a bool
value indicating that.

Modify acpi_evaluate_reference() as per the observation above and update
its callers accordingly so as to get rid of useless code and local
variables.

The observable behavior of the kernel is not expected to change after
this modification of the code.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

K2CI-Arch: All
Change-Id: I03e1ecb53316a5c3000ecfbd6c9c8df5b26ed1d0
Signed-off-by: Yufeng Wang <wangyufeng@kylinos.cn>
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/146512
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>