arm64: mm: Move fixmap region above vmemmap region
Mainline: b730b0f2b1fcfbdaed816152cc71993fd708aa11
From: v6.9-rc1
Severity: Low

Move the fixmap region above the vmemmap region, so that the start of
the vmemmap delineates the end of the region available for vmalloc and
vmap allocations and the randomized placement of the kernel and modules.

In a subsequent patch, we will take advantage of this to reclaim most of
the vmemmap area when running a 52-bit VA capable build with 52-bit
virtual addressing disabled at runtime.

Note that the existing guard region of 256 MiB covers the fixmap and PCI
I/O regions as well, so we can reduce it 8 MiB, which is what we use in
other places too.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Link: https://lore.kernel.org/r/20231213084024.2367360-11-ardb@google.com
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Acked-by: Mark Rutland <mark.rutland@arm.com>

(backported from commit b730b0f2b1fcfbdaed816152cc71993fd708aa11)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I9b436d0a569f61e3faae160f6f15bbffa067b1b2
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/160540
Tested-by: k2ci <kernel-bot@kylinos.cn>