arm64: kaslr: Adjust randomization range dynamically
Mainline: 3567fa63cb5680d3e1e8375c547a0e305c8a0ff5
From: v6.9-rc1
Severity: Low

Currently, we base the KASLR randomization range on a rough estimate of
the available space in the upper VA region: the lower 1/4th has the
module region and the upper 1/4th has the fixmap, vmemmap and PCI I/O
ranges, and so we pick a random location in the remaining space in the
middle.

Once we enable support for 5-level paging with 4k pages, this no longer
works: the vmemmap region, being dimensioned to cover a 52-bit linear
region, takes up so much space in the upper VA region (the size of which
is based on a 48-bit VA space for compatibility with non-LVA hardware)
that the region above the vmalloc region takes up more than a quarter of
the available space.

So instead of a heuristic, let's derive the randomization range from the
actual boundaries of the vmalloc region.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Link: https://lore.kernel.org/r/20231213084024.2367360-16-ardb@google.com
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Acked-by: Mark Rutland <mark.rutland@arm.com>

(backported from commit 3567fa63cb5680d3e1e8375c547a0e305c8a0ff5)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I3c89e0a84cd0136bc88b61d5574ee3fde43d0a42
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/160545
Tested-by: k2ci <kernel-bot@kylinos.cn>