ext4: fix inconsistent between segment fstrim and full fstrim
Mainline: 68da4c44b994aea797eb9821acb3a4a36015293e
From: v6.8-rc1
Severity: Moderate

task: 198520

[ Upstream commit 68da4c44b994aea797eb9821acb3a4a36015293e ]

Suppose we issue two FITRIM ioctls for ranges [0,15] and [16,31] with
mininum length of trimmed range set to 8 blocks. If we have say a range of
blocks 10-22 free, this range will not be trimmed because it straddles the
boundary of the two FITRIM ranges and neither part is big enough. This is a
bit surprising to some users that call FITRIM on smaller ranges of blocks
to limit impact on the system. Also XFS trims all free space extents that
overlap with the specified range so we are inconsistent among filesystems.
Let's change ext4_try_to_trim_range() to consider for trimming the whole
free space extent that straddles the end of specified range, not just the
part of it within the range.

Signed-off-by: Ye Bin <yebin10@huawei.com>
Reviewed-by: Jan Kara <jack@suse.cz>
Link: https://lore.kernel.org/r/20231216010919.1995851-1-yebin10@huawei.com
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit c1eacba3b77cfa808c06b2c40d8ecab5d4f69f00)

K2CI-Arch: None
Signed-off-by: k2ci <kernel-bot@kylinos.cn>
Change-Id: I0d14a5bb6d98795b70ea11f189c544b84e90a8c4
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/96540
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>