driver core: bus: constantify subsys_register() calls
Mainline: 32f78abe59c740b6ec34c89dc10a09208eae7e1f
From: v6.8-rc1
Severity: Low

task: #482381

The functions subsys_register() and subsys_virtual_register() should be
taking a constant pointer to a struct bus_type, as they do not actually
modify anything in it, so fix up the function definitions to do so
properly.

This also changes the pointer type in struct subsys_interface to be
constant as well, as again, that's the proper signature of it.

Cc: Rafael J. Wysocki <rafael@kernel.org>
Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Link: https://lore.kernel.org/r/2023121908-grove-genetics-f8af@gregkh
Change-Id: I5a885c6bbac9add11858605f5fae64becd6a6329
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
K2CI-Arch: All
Signed-off-by: zhongkaihua <zhongkaihua@kylinos.cn>
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/80231
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 刘云 <liuyun01@kylinos.cn>