wifi: mac80211: Skip association timeout update after comeback rejection
Mainline: b133fdf07db849eb2ce25df19d88ff31701d3b6d
From: v6.8-rc1
Severity: Moderate

task: 496179

It is possible for the TX status report for the (Re)Association Request
frame to be delayed long enough for the AP's (Re)Association Response
frame to be received and processed before it. If that were to happen for
a case where the AP rejects the association with indication to come back
later, the association timeout and retry state should not be modified
anymore with the TX status information that would be processed after
this. Updating the association timeout in such a reverse order of events
could result in shortening the timeouts for the association comeback
mechanism and that could result in the association failing.

Track whether we have already processed association rejection with
comeback time and if so, skip the timeout and retry update on any
following TX status report.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Link: https://msgid.link/20231219174814.2581575-1-j@w1.fi
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

(backported from commit b133fdf07db849eb2ce25df19d88ff31701d3b6d)

K2CI-Arch: All
Change-Id: I1d812ca44ad64495041e2f1070b3acfca9018fe0
Signed-off-by: Yi Cong <yicong@kylinos.cn>
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/91878
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 王琦 <wangqi@kylinos.cn>