OPENEULER: RDMA/hns: Fix cpu stuck by printings during reset
Mainline: KYLIN-only
Severity: Moderate

Task: #219093

During the reset, issuing the cmdq command to destroy resources
such as qp, cq, and mr may fail, and the destruction failure
information is printed. When a large number of resources fail to
be destroyed, printk() causes multiple threads to compete for a
global log buffer spinlock.  As a result, the CPU performance
deteriorates and the CPU is stuck. To fix it, use a function like
printk_limit_rate() to limit the print frequency of printk().

Fixes: 9a4435375cd1 ("IB/hns: Add driver files for hns RoCE driver")
Fixes: c7bcb13442e1 ("RDMA/hns: Add SRQ support for hip08 kernel mode")
Fixes: 626903e9355b ("RDMA/hns: Add support for reporting wc as software mode")
Fixes: 6eef524201de ("RDMA/hns: Replace not intuitive function/macro names")
Fixes: 357f34294686 ("RDMA/hns: Simplify the state judgment code of qp")
Fixes: 926a01dc000d ("RDMA/hns: Add QP operations support for hip08 SoC")
Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>
Signed-off-by: Juan Zhou <zhoujuan51@h-partners.com>

K2CI-Arch: All
Link: https://gitee.com/openeuler/kernel/issues/I9FIHP
Signed-off-by: Yongzhen Zhang <zhangyongzhen@kylinos.cn>
Change-Id: I475e23f3d83e437e4268f3731fcfd2ca39ce9cba
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/113658
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>