wifi: cfg80211: free beacon_ies when overridden from hidden BSS
Mainline: 32af9a9e1069e55bc02741fb00ac9d0ca1a2eaef
From: v6.8-rc1
Severity: Moderate

task: 198520

[ Upstream commit 32af9a9e1069e55bc02741fb00ac9d0ca1a2eaef ]

This is a more of a cosmetic fix. The branch will only be taken if
proberesp_ies is set, which implies that beacon_ies is not set unless we
are connected to an AP that just did a channel switch. And, in that case
we should have found the BSS in the internal storage to begin with.

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Link: https://msgid.link/20231220133549.b898e22dadff.Id8c4c10aedd176ef2e18a4cad747b299f150f9df@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>

(backported from commit 0dc97f76d67079969bf7f16ab0863bd304268b17)

K2CI-Arch: None
Signed-off-by: k2ci <kernel-bot@kylinos.cn>
Change-Id: Id9fffdf4f0f188ca0697808f600e005d70ddb5d2
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/96629
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>