RISC-V: KVM: Add steal-update vcpu request
Mainline: 2a1f6bf079700f0f9d8045ab77b302aeb4d12c06
From: v6.8-rc1
Severity: Low

Task: #551019

Add a new vcpu request to inform a vcpu that it should record its
steal-time information. The request is made each time it has been
detected that the vcpu task was not assigned a cpu for some time,
which is easy to do by making the request from vcpu-load. The record
function is just a stub for now and will be filled in with the rest
of the steal-time support functions in following patches.

Reviewed-by: Anup Patel <anup@brainfault.org>
Reviewed-by: Atish Patra <atishp@rivosinc.com>
Signed-off-by: Andrew Jones <ajones@ventanamicro.com>
Signed-off-by: Anup Patel <anup@brainfault.org>

K2CI-Arch: All
Change-Id: Ie0ea215c7283ce4b2a0c2faf5271dd3a9c9cd404
Signed-off-by: Yufeng Wang <wangyufeng@kylinos.cn>
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/148650
Reviewed-by: 李唯实 <liweishi@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>