OPENEULER: roh/hns3: Fix the processing flow of ROH CMDq during the reset process.
Mainline: KYLIN-only
Severity: Moderate

task: 219056

Fix the sequence issue in the ROH CMDq processing flow during
the FLR reset process, which may occasionally cause calltrace.
In this patch, we added a flag in the reset process to indicate
the availability of the ROH CMDq, in order to fix this issue.

Fixes: 1cb9fff33035 ("roh/hns3: Add ROH cmdq interface support")
Signed-off-by: Ke Chen <chenke54@huawei.com>
Reviewed-by: Gang Zhang <gang.zhang@huawei.com>

K2CI-Arch: Arm64
Link: https://gitee.com/openeuler/kernel/issues/I9C19A
Signed-off-by: lifangxu <lifangxu@kylinos.cn>
Change-Id: Ic765e469fc3524702fd765b160432762fe51718b
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/115772
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>