OPENEULER: livepatch/core: Restrict livepatch patched/unpatched when plant kprobe
Mainline: KYLIN-only
Severity: Low

task: #205708

livepatch wo_ftrace and kprobe are in conflict, because kprobe
may modify the instructions anywhere in the function.

So it's dangerous to patched/unpatched an function when there are
some kprobes registered on it. Restrict these situation.

we should hold kprobe_mutex in klp_check_patch_kprobed, but it's
static and can't export, so protect klp_check_patch_probe in
stop_machine to avoid registing kprobes when patching.

we do nothing for (un)register kprobes on the (old) function
which has been patched. because there are sone engineers need this.
certainly, it will not lead to hangs, but not recommended.

Signed-off-by: Cheng Jian <cj.chengjian@huawei.com>
Signed-off-by: Wang ShaoBo <bobo.shaobowang@huawei.com>
Signed-off-by: Dong Kai <dongkai11@huawei.com>
Signed-off-by: Ye Weihua <yeweihua4@huawei.com>
Signed-off-by: Zheng Yejian <zhengyejian1@huawei.com>

K2CI-Arch: Arm64|Amd64
Link: https://gitee.com/openeuler/kernel/issues/I8MGE6
Signed-off-by: zhangxi <zhangxi@kylinos.cn>
Change-Id: I1a77b381ce82972117497f1211b5b24bc2f9f024
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/114571
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Reviewed-by: 郭东太 <guodongtai@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>