OPENEULER: livepatch: Use breakpoint exception to optimize enabling livepatch
Mainline: KYLIN-only
Severity: Low

task: #205708

The commit 86e35fae15bb ("livepatch: checks only if the replaced
instruction is on the stack") optimizes stack checking. However, for
extremely hot functions, the replaced instruction may still be on the
stack, and there is room for further optimization.

By inserting a breakpoint exception instruction at the entry of the
patched old function, we can divert calls from the old function to the
new function. In this way, during stack check, only tasks that have
entered the old function before the breakpoint is inserted need to be
considered. This increases the probability of passing the stack check.

If the stack check fails, we sleep for a period of time and try again,
giving the task entering the old function a chance to run out of the
instruction replacement area.

We first enable the patch using the normal process, that is, do not
insert breakpoints. If the first enable fails and the force flag
KLP_STACK_OPTIMIZE is set for all functions of the patch, then we use
breakpoint exception optimization.

Signed-off-by: Li Huafei <lihuafei1@huawei.com>
Signed-off-by: Zheng Yejian <zhengyejian1@huawei.com>

K2CI-Arch: Arm64|Amd64
Link: https://gitee.com/openeuler/kernel/issues/I8MGE6
Signed-off-by: zhangxi <zhangxi@kylinos.cn>
Change-Id: Ia979d91de01691695dedce111ce8d932b7090dec
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/114581
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Reviewed-by: 郭东太 <guodongtai@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>